Fetch Historical Exchange Rates from Yahoo Finance

This function retrieves historical exchange rates for a given currency pair from Yahoo Finance and returns a dataframe with daily exchange rate data (values are supposed to be each precise minute of the same day).

Usage

fetch_changes(from, to)

Arguments

from

A character string representing the base currency (e.g., "USD").

to

A character string representing the target currency (e.g., "EUR").

Value

A dataframe with daily exchange rates. The returned dataframe contains the following columns:

timestamp

int, the timestamp of the rate.

rate.high

Numeric, the highest exchange rate of the day.

rate.low

Numeric, the lowest exchange rate of the day.

rate.open

Numeric, the opening exchange rate of the day.

rate.volume

Numeric, the trading volume for the day.

rate.close

Numeric, the closing exchange rate of the day.

date

POSIXct, the corresponding date (YYYY-MM-DD).

Details

The function also returns an additional insights data.frame as attributes. Some of the key attributes include latest informations from Yahoo Finance API :

currency

Character, the base currency.

symbol

Character, the Yahoo Finance symbol (e.g., "EURUSD=X").

exchange_name

Character, the exchange place name.

regular_market_price

Numeric, the latest market price.

regularmarketdayhigh

Numeric, the market highest price of this day.

regularmarketdaylow

Numeric, the market lowest price of this day.

fifty_two_week_high

Numeric, the highest price in the last 52 weeks.

fifty_two_week_low

Numeric, the lowest price in the last 52 weeks.

previous_close

Numeric, the last closing price.

timezone

Character, the market's timezone.

range

Character, the data range (e.g., "1d", "5d", "1mo").

Examples

# Fetch exchange rate data for USD to EUR
df <- fetch_changes("USD", "EUR")

# View the first few rows of the dataframe
head(df)
#>    timestamp rate.close rate.volume rate.open rate.high rate.low
#> 1 1742860800     0.9255           0    0.9255    0.9255   0.9255
#> 2 1742860860     0.9256           0    0.9256    0.9256   0.9256
#> 3 1742860920     0.9258           0    0.9258    0.9258   0.9258
#> 4 1742860980     0.9257           0    0.9257    0.9257   0.9257
#> 5 1742861040     0.9257           0    0.9257    0.9257   0.9257
#> 6 1742861100     0.9257           0    0.9257    0.9257   0.9257
#>                  date
#> 1 2025-03-25 00:00:00
#> 2 2025-03-25 00:01:00
#> 3 2025-03-25 00:02:00
#> 4 2025-03-25 00:03:00
#> 5 2025-03-25 00:04:00
#> 6 2025-03-25 00:05:00

# Access metadata attributes
attr(df, "insights")
#>   currency   symbol exchangename fullexchangename instrumenttype firsttradedate
#> 1      EUR USDEUR=X          CCY              CCY       CURRENCY     1070236800
#>   regularmarkettime hasprepostmarketdata gmtoffset timezone
#> 1        1742946747                FALSE         0      GMT
#>   exchangetimezonename regularmarketprice fiftytwoweekhigh fiftytwoweeklow
#> 1        Europe/London             0.9268           0.9819          0.8919
#>   regularmarketdayhigh regularmarketdaylow regularmarketvolume longname
#> 1               0.9269              0.9262                   0  USD/EUR
#>   shortname chartpreviousclose previousclose scale pricehint
#> 1   USD/EUR             0.9256        0.9256     4         4
#>                   tradingperiods datagranularity range
#> 1 GMT, 1742947140, 1742860800, 0              1m    1d
#>                                        validranges pre.timezone    pre.end
#> 1 1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max          GMT 1742860800
#>    pre.start pre.gmtoffset regular.timezone regular.end regular.start
#> 1 1742860800             0              GMT  1742947140    1742860800
#>   regular.gmtoffset post.timezone   post.end post.start post.gmtoffset
#> 1                 0           GMT 1742947140 1742947140              0