Fetch stock indices with quantmod

Given indices names, the function retrieves historical stock market data (from finance.yahoo.com by default). It returns the data frame returned by quantmod::getSymbols with a proper date column, standardized colnames.

fetch_historic(
  symbols = c("SAAB-B.ST"),
  wait.time = 0,
  .verbose = T,
  src = "yahoo"
)

Arguments

symbols

character A character string representing the indices to search for.

wait.time

double A character string representing the indices to search for.

.verbose

logical, default = TRUE. If TRUE, send messages to the console

src

character, default = 'yahoo' The src parameters of quantmod::getSymbols.

Value

A data frame with columns: - open: The value at the opening of the marketplace - high: Highest value of the indice during this day. - low: Lowest value of the indice during this day. - close: Value at the closing of the maketplace. - volume: The volume of operation. - adjusted: The adjusted closing value.

symbol: The original indices names searched

date: The date for these values

See also

Examples

datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
#> ETA :  0.5  sec'