Fetch stock indices based on a company name

Given companies names, the function retrieves overall stock market data (by searching on https://finance.yahoo.com ). It returns a data frame with the stock symbol on various marketplaces , companies names, last price on the marketplace, sector/category (if available), type (always "stocks") , exchange marketplace name and initially searched companies names.

fetch_stock_indices(names, marketplaces = NULL)

Arguments

names

A character string representing the company name to search for.

marketplaces

(optionnal) A character string representing the marketplace(s) to consider. Default keep all the marketplace

Value

A data frame with columns: - Symbol: The stock ticker symbol from yahoo - Name: The full company name. - Last Price: The latest available price. - Sector / Category: The sector or industry category (if available). - Type: The type of asset (always "stocks"). - Exchange: The stock exchange place for this stock.

searched: The original names searched

Examples

fetch_stock_indices(names = c("VOLVO", "SAAB"),  marketplaces = "STO"  )
#>         symbol                name last price sector / category   type exchange
#> 1  VOLCAR-B.ST Volvo Car AB ser. B      21.89 Consumer Cyclical Stocks      STO
#> 4    VOLV-A.ST    Volvo, AB ser. A     308.80       Industrials Stocks      STO
#> 6    VOLV-B.ST    Volvo, AB ser. B     308.90       Industrials Stocks      STO
#> 26   SAAB-B.ST      SAAB AB ser. B     380.05       Industrials Stocks      STO
#>    searched
#> 1     VOLVO
#> 4     VOLVO
#> 6     VOLVO
#> 26     SAAB