Get Exchange Rates from Yahoo Finance, Given Devises

This function retrieves exchange rates for a given currency pair. Default convert to USD ($) Source is Yahoo Finance API. It returns a data.frame with daily exchange rate data. Returned data.frame is historical values on a day and latest insights within attributes Returned values are supposed to be each minute within a day, more or less (≃ 1200 obs. ± 200 obs.).

Usage

get_changes(from = NULL, to = "USD")

Arguments

from,

default = NULL A character string representing the base currency (e.g., "USD").

to,

default = "USD" A character string representing the target currency (e.g., "EUR").

Value

A dataframe with daily exchange rates. The returned dataframe contains the following columns:

timestamp

int, the timestamp of the rate.

rate.high

Numeric, the highest exchange rate of the day.

rate.low

Numeric, the lowest exchange rate of the day.

rate.open

Numeric, the opening exchange rate of the day.

rate.volume

Numeric, the trading volume for the day.

rate.close

Numeric, the closing exchange rate of the day.

date

POSIXct, the corresponding date (YYYY-MM-DD).

Details

The function also returns an additional insights data.frame as attributes. Some of the key attributes include latest informations from Yahoo Finance API :

currency

character, the base currency.

symbol

character, the Yahoo Finance symbol (e.g., "EURUSD=X").

exchange_name

character, the exchange place name, i.e. CCY for currencies.

regularmarketprice

numeric, the latest market price.

regularmarketdayhigh

numeric, the market highest price of this day.

regularmarketdaylow

numeric, the market lowest price of this day.

fifty_two_week_high

numeric, the highest price in the last 52 weeks.

fifty_two_week_low

numeric, the lowest price in the last 52 weeks.

previous_close

numeric, the last closing price.

timezone

character, the market's timezone.

date

POSIXct Equivalent to the timestamp of regularmarketprice, date of the observation.

from

character, the currency converted into another, e.g., if the from value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.

to

character, the currency that you want to convert into : all the numeric values (not integer) in this line of the data.frame are expressed with this currency.

Examples

# Fetch exchange rate data for USD to EUR
df <- get_changes(from = c("USD", "EUR"), to = "RON")
# Convert USD & EUR to Romanian Leu (RON)
head(df)
#>   currency   symbol exchangename fullexchangename instrumenttype firsttradedate
#> 1      RON USDRON=X          CCY              CCY       CURRENCY     1120172400
#> 2          EURRON=X          CCY              CCY       CURRENCY     1120172400
#>   regularmarkettime hasprepostmarketdata gmtoffset timezone
#> 1        1743119924                FALSE         0      GMT
#> 2        1743119725                FALSE         0      GMT
#>   exchangetimezonename regularmarketprice fiftytwoweekhigh fiftytwoweeklow
#> 1        Europe/London             4.6060           4.8828          4.4281
#> 2        Europe/London             4.9767           4.9820          4.9060
#>   regularmarketdayhigh regularmarketdaylow regularmarketvolume longname
#> 1               4.6083              4.6058                   0  USD/RON
#> 2               4.9776              4.9761                   0  EUR/RON
#>   shortname chartpreviousclose previousclose scale pricehint datagranularity
#> 1   USD/RON             4.6310        4.6310     4         4              1m
#> 2   EUR/RON             4.9753        4.9753     4         4              1m
#>   range pre_timezone    pre_end  pre_start pre_gmtoffset regular_timezone
#> 1    1d          GMT 1743033600 1743033600             0              GMT
#> 2    1d          GMT 1743033600 1743033600             0              GMT
#>   regular_end regular_start regular_gmtoffset post_timezone   post_end
#> 1  1743119940    1743033600                 0           GMT 1743119940
#> 2  1743119940    1743033600                 0           GMT 1743119940
#>   post_start post_gmtoffset                date from  to
#> 1 1743119940              0 2025-03-27 23:58:44  USD RON
#> 2 1743119940              0 2025-03-27 23:55:25  EUR RON
# or pass like a named list of character
df2 <- get_changes(from = c("EUR" = "RON", "USD" = "EUR"))
# Or indicate paired values as 2 list :
same_as_df2 <- get_changes(from = c("EUR", "USD"), c("RON" , "EUR"))
# Access metadata attributes
attr(df, "insights")
#> NULL