Fetch historical financial data indices

Given indices names (i.e. ticker symbol), the function retrieves historical stock market data from finance.yahoo.com. Answer a data.frame with standardized colnames.

Usage

get_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T, ...)

Arguments

symbols

character A character string representing the financial indices to search for, e.g., ticker symbol(s).

wait.time

double, default = 0 A character string representing an additional waiting time between 2 calls to the Yahoo API.

.verbose

logical, default = TRUE. If TRUE, send messages to the console.

...

Arguments passed on to get_yahoo_data

symbol

A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).

start_date

A character string representing the start date in the format "YYYY-MM-DD". If NULL, data starts from 1970-01-01.

end_date

A character string representing the end date in the format "YYYY-MM-DD". If NULL, data is retrieved up to the current date.

interval

A character string representing the interval for the returned datas : default will return daily values. Other intervals will filter out some of these daily values, depending on the desired interval. Valid interval are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".

Value

A data frame containing the historical financial data with the following columns:

open

numeric The opening price for the period (default is each day).

close

numeric The closing price for the period (default is each day).

adjclose

numeric The adjusted closing price on the period, which accounts for corporate actions like dividends and stock splits.

low

numeric The lowest price for the period (default is each day).

high

numeric The highest price for the period (default is each day).

volume

integer The traded volume.

timestamp

integer Unix timestamps corresponding to each data point.

date

POSIXct The day of the financial data point.

currency

character The currency in which the data is reported, depending on the marketplace.

symbol

character The stock or financial instrument symbol (e.g., "AAPL").

shortname

character The abbreviated name of the company or financial instrument.

longname

character The full name of the company or financial instrument.

exchangename

character The name of the exchange marketplace where the financial instrument is listed.

fullexchangename

character The full name of the exchange marketplace.

timezone

character The timezone in which the data is reported.

Details

Return a data.frame (see returned columns in 'Value' section) with additional attributes :

symbols: the symbols originally asked by the user

date: the date when data are retrieved

currencies: the currencies within the data.frame

n.currencies: the number of currencies within the data.frame

min.date: lowest (ancient) date-value within the data.frame

max.date: maximum (recent) date-value within the data.frame

See also

Examples

datas <- get_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
#> 
#> => 2 request(s) to Yahoo Finance (ETA :  0.6  sec')  VOLCAR-B.ST [OK]                                                                                                      SAAB-B.ST [OK]