Fetch stock indices based on a company name

Given companies names, the function retrieves overall stock market data (by searching on https://finance.yahoo.com ). It returns a data frame with the stock symbol on various marketplaces , companies names, last price on the marketplace, sector/category (if available), type (always "stocks") , exchange marketplace name and initially searched companies names.

fetch_stock_indices(names)

Arguments

names

A character string representing the company name to search for.

Value

A data frame with columns: - Symbol: The stock ticker symbol from yahoo - Name: The full company name. - Last Price: The latest available price. - Sector / Category: The sector or industry category (if available). - Type: The type of asset (always "stocks"). - Exchange: The stock exchange place for this stock.

searched: The original names searched

Examples

fetch_stock_indices(names = c("VOLVO", "SAAB") )
#>       Symbol                            Name   Last Price   Sector / Category  
#> 1  VOLCAR-B.ST             Volvo Car AB ser. B        21.89   Consumer Cyclical
#> 2      VOL1.SG                           Volvo        28.06         Industrials
#> 3      VOL1.MU Volvo (publ), AB              N        28.41         Industrials
#> 4      VOL3.BE Volvo (publ), AB              N        28.26         Industrials
#> 5        VLVLY                        Volvo AB        30.39         Industrials
#> 6    VOLVBS.XD                        Volvo AB       308.55                  --
#> 7      VOL1.DU Volvo (publ), AB              N        28.02         Industrials
#> 8    VOLCBS.XC                    Volvo Car AB        21.83   Consumer Cyclical
#> 9     VOLVB.PR                        VOLVO AB       760.00         Industrials
#> 10   VOLVBS.XC                        Volvo AB       309.60                  --
#> 11     VOL3.HA Volvo (publ), AB              N        28.26         Industrials
#> 12   VOLV-B.ST                Volvo, AB ser. B       308.90         Industrials
#> 13   VOLV-A.ST                Volvo, AB ser. A       308.80         Industrials
#> 14     VOL3.MU Volvo (publ), AB              N        28.24         Industrials
#> 15     0HTP.IL  VOLVO AB VOLVO CLASS B ORD SHS       309.25         Industrials
#> 16     VOL1.BE Volvo (publ), AB              N        28.20         Industrials
#> 17       VOLVF                        Volvo AB        31.74         Industrials
#> 18   VOLCBS.XD                    Volvo Car AB        21.91                  --
#> 19   VOLVAS.XD                        Volvo AB       308.80                  --
#> 20     0MHW.IL  VOLVO AB VOLVO CLASS A ORD SHS       309.30         Industrials
#> 21      VOL3.F Volvo (publ), AB              N        28.26         Industrials
#> 22     VOL3.SG                   AB Volvo publ        28.08         Industrials
#> 23     0AAK.IL  VOLVO CAR AB VOLVO CAR ORD SHS        21.83   Consumer Cyclical
#> 24      VOL4.F Volvo (publ), AB              N        28.00         Industrials
#> 25     VOL3.HM Volvo (publ), AB              N        28.26         Industrials
#> 26   SAAB-B.ST                  SAAB AB ser. B       380.05         Industrials
#> 27       SAABY                         Saab AB        18.80         Industrials
#> 28      SDV1.F Saab AB                       N        34.94         Industrials
#> 29       SAABF                         Saab AB        37.65         Industrials
#> 30     0GWL.IL          SAAB AB SAAB B ORD SHS       379.45         Industrials
#> 31   SAABBS.XC                         Saab AB       361.95         Industrials
#> 32   SAABBS.XD                         Saab AB       381.30                  --
#> 33     SDV0.MU Saab AB                       N        18.00         Industrials
#> 34     SDV1.SG                         Saab AB        34.78         Industrials
#> 35     SDV0.BE Saab AB                       N        16.70         Industrials
#> 36      SDV0.F Saab AB                       N        16.60         Industrials
#> 37     SDV1.DU Saab AB                       N        34.68         Industrials
#> 38     SDV0.SG                  Saab AB (ADRs)        16.50         Industrials
#> 39     SDV1.MU Saab AB                       N        34.67         Industrials
#> 40     SDV1.BE Saab AB                       N        34.76         Industrials
#>    Type   Exchange   searched
#> 1  Stocks        STO    VOLVO
#> 2  Stocks        STU    VOLVO
#> 3  Stocks        MUN    VOLVO
#> 4  Stocks        BER    VOLVO
#> 5  Stocks        PNK    VOLVO
#> 6  Stocks        DXE    VOLVO
#> 7  Stocks        DUS    VOLVO
#> 8  Stocks        CXE    VOLVO
#> 9  Stocks        PRA    VOLVO
#> 10 Stocks        CXE    VOLVO
#> 11 Stocks        HAN    VOLVO
#> 12 Stocks        STO    VOLVO
#> 13 Stocks        STO    VOLVO
#> 14 Stocks        MUN    VOLVO
#> 15 Stocks        IOB    VOLVO
#> 16 Stocks        BER    VOLVO
#> 17 Stocks        PNK    VOLVO
#> 18 Stocks        DXE    VOLVO
#> 19 Stocks        DXE    VOLVO
#> 20 Stocks        IOB    VOLVO
#> 21 Stocks        FRA    VOLVO
#> 22 Stocks        STU    VOLVO
#> 23 Stocks        IOB    VOLVO
#> 24 Stocks        FRA    VOLVO
#> 25 Stocks        HAM    VOLVO
#> 26 Stocks        STO     SAAB
#> 27 Stocks        PNK     SAAB
#> 28 Stocks        FRA     SAAB
#> 29 Stocks        PNK     SAAB
#> 30 Stocks        IOB     SAAB
#> 31 Stocks        CXE     SAAB
#> 32 Stocks        DXE     SAAB
#> 33 Stocks        MUN     SAAB
#> 34 Stocks        STU     SAAB
#> 35 Stocks        BER     SAAB
#> 36 Stocks        FRA     SAAB
#> 37 Stocks        DUS     SAAB
#> 38 Stocks        STU     SAAB
#> 39 Stocks        MUN     SAAB
#> 40 Stocks        BER     SAAB