Get Yahoo Finance Data for a Given Symbol

Given the symbol of a financial value, this function retrieves historical financial data from Yahoo Finance and answer a data.frame of historical values. The data includes open stock price, high, low, close, volume, and adjusted close prices, along with timestamps (i.e. the day associated with the measures).

Usage

get_yahoo_data(
  symbol = "AAPL",
  start_date = NULL,
  end_date = NULL,
  interval = "1d"
)

Arguments

symbol

A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).

start_date

A character string representing the start date in the format "YYYY-MM-DD". If NULL, data starts from 1970-01-01.

end_date

A character string representing the end date in the format "YYYY-MM-DD". If NULL, data is retrieved up to the current date.

interval

A character string representing the interval for the returned datas : default will return daily values. Other intervals will filter out some of these daily values, depending on the desired interval. Valid interval are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".

Value

A data frame containing the historical financial data with the following columns:

volume

integer The traded volume.

high

numeric The highest price for the period (default is each day).

open

numeric The opening price for the period (default is each day).

low

numeric The lowest price for the period (default is each day).

close

numeric The closing price for the period (default is each day).

adjclose

numeric The adjusted closing price on the period, which accounts for corporate actions like dividends and stock splits.

timestamp

integer Unix timestamps corresponding to each data point.

date

POSIXct The day of the financial data point.

currency

The currency in which the data is reported, depending on the marketplace.

symbol

The stock or financial instrument symbol (e.g., "AAPL").

longname

The full name of the company or financial instrument.

shortname

The abbreviated name of the company or financial instrument.

exchangename

The name of the exchange marketplace where the financial instrument is listed.

fullexchangename

The full name of the exchange marketplace.

timezone

The timezone in which the data is reported.

Details

The default data.frame have a line for each day. If the user provide another interval range than '1d' (one day), lines will be filtered out, in order to match the desired interval range. Valid interval ranges are: "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".

The function allows the user to specify a date range using start and end dates. If no date range is specified, it retrieves all available data from the beginning of time (default for start) to the current date (default for end).

Examples

  data <- get_yahoo_data(symbol = "SAAB-B.ST", start_date = "2020-01-01", interval = "5d")
  head(data)#'
#>     open  close  volume   high    low adjclose  timestamp                date
#> 1 84.575 85.050 1671108 85.050 83.825 80.94995 1579161600 2020-01-16 08:00:00
#> 2 86.000 85.300  888888 86.000 84.750 81.18790 1579593600 2020-01-21 08:00:00
#> 3 78.825 78.300 1271576 79.350 77.925 74.52535 1580457600 2020-01-31 08:00:00
#> 4 81.750 83.300 1503696 83.400 81.300 79.28431 1580889600 2020-02-05 08:00:00
#> 5 79.800 77.500 2685416 80.025 77.200 73.76392 1581321600 2020-02-10 08:00:00
#> 6 74.500 73.925 1220424 75.000 73.900 70.36127 1582185600 2020-02-20 08:00:00
#>   currency    symbol       longname      shortname exchangename
#> 1      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#> 2      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#> 3      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#> 4      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#> 5      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#> 6      SEK SAAB-B.ST Saab AB (publ) SAAB AB ser. B          STO
#>   fullexchangename timezone
#> 1        Stockholm      CET
#> 2        Stockholm      CET
#> 3        Stockholm      CET
#> 4        Stockholm      CET
#> 5        Stockholm      CET
#> 6        Stockholm      CET
  invalid_datas <- get_yahoo_data(symbol ="SAAB" )
#> Warning: The value is not associated with a valid currency name !
#> => SAAB is not a valid currency name, according to the Yahoo Finance API.
#> See : https://finance.yahoo.com/ or stockr::fetch_stock_indices
  invalid_datas
#> [1] NA